+++ src/drivers/lcd/pcd8544/displayPcd8544.hpp
@@ -22,9 +22,9 @@ public:
     DisplayPcd8544(bsp::Board& board);
 
     void display() override;
-    u16 getHeight() const override;
-    u16 getWidth() const override;
-    void setPixel(u16 x, u16 y, display::Colors color = display::Colors::BLACK) override;
+    int getHeight() const override;
+    int getWidth() const override;
+    void setPixel(int x, int y, display::Colors color = display::Colors::BLACK) override;
 
 protected:
     void reset();
