+++ src/drivers/lcd/sfml/displaySfml.hpp
@@ -17,9 +17,9 @@ public:
     DisplaySfml();
 
     void display() override;
-    u16 getHeight() const override;
-    u16 getWidth() const override;
-    void setPixel(u16 x, u16 y, display::Colors color = display::Colors::BLACK) override;
+    int getHeight() const override;
+    int getWidth() const override;
+    void setPixel(int x, int y, display::Colors color = display::Colors::BLACK) override;
 
 protected:
     static sf::Color convertToSfColor(display::Colors color);
